Separate GenericDeserializer.cs
