+++ Assets/VRM/UniJSON/Scripts/GenericDeserializer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 74625bf2a3d6bc64d9d92b08221a8a3d
+timeCreated: 1549450914
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
